refactor: move `providers/{openai|anthropic}` out of the `providers/` dir
