+++ Taskfile.yaml
@@ -6,9 +6,9 @@ env:
 vars:
   PACKAGES:
     - ./ai
+    - ./anthropic
     - ./examples
-    - ./providers/anthropic
-    - ./providers/openai
+    - ./openai
 
 tasks:
   fmt:
