+++ examples/streaming-agent/main.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/charmbracelet/ai/ai"
-	"github.com/charmbracelet/ai/providers/anthropic"
+	"github.com/charmbracelet/ai/anthropic"
 )
 
 func main() {
