+++ go.work
@@ -2,7 +2,7 @@ go 1.24.5
 
 use (
 	./ai
+	./anthropic
 	./examples
-	./providers/anthropic
-	./providers/openai
+	./openai
 )
