+++ openai/json.go
