+++ openai/openai.go
